Add VaultModel
- Encapsulates the small logic for this code. Allows it to be used by Commands, which need to be defined up for the Scene, which is not in this SPM.
