+++ Sources/Site/Music/UI/VaultView.swift
@@ -13,44 +13,28 @@ extension Logger {
 }
 
 public struct VaultView: View {
-  let url: URL
+  @ObservedObject private var model: VaultModel
 
-  @State private var vault: Vault? = nil
-  @State private var error: Error? = nil
-
-  public init(url: URL) {
-    self.url = url
-  }
-
-  private func loadVault() async {
-    Logger.vaultLoad.log("start")
-    defer {
-      Logger.vaultLoad.log("end")
-    }
-    do {
-      vault = try await Vault.load(url: url)
-    } catch {
-      Logger.vaultLoad.log("error: \(error.localizedDescription, privacy: .public)")
-      self.error = error
-    }
+  public init(_ model: VaultModel) {
+    self.model = model
   }
 
   public var body: some View {
     Group {
-      if let vault {
+      if let vault = model.vault {
         ArchiveCategorySplit(vault: vault)
           .refreshable {
             Logger.vaultLoad.log("refresh")
-            await loadVault()
+            await model.load()
           }
-      } else if let error {
+      } else if let error = model.error {
         Text(error.localizedDescription)
       } else {
         ProgressView()
       }
     }.task {
       Logger.vaultLoad.log("task")
-      await loadVault()
+      await model.load()
     }
   }
 }
